+++ spec/factories/api_keys.rb
@@ -24,4 +24,12 @@ FactoryGirl.define do
     name { unique_name('internal api key') }
     grants [{ type: "apis", apis: [] }]
   end
+
+  factory :oauth_api_key_user_profile_grant, class: Carto::ApiKey do
+    initialize_with { Carto::ApiKey.send :new }
+
+    type Carto::ApiKey::TYPE_OAUTH
+    name { unique_name('internal api key') }
+    grants [{ type: "apis", apis: [] }, { type: 'user', data: ['profile'] }]
+  end
 end